renamed and merged phreeqc_cpp 8037
git-svn-id: svn://136.177.114.72/svn_GW/phastpp/branches/phastpp_class_mpi@8038 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ NNInterpolator/nnpi.c
@@ -833,7 +833,7 @@ nnpi_interpolate_points(int nin, point pin[], double wmin, int nout,
 		{
 			point *p = &pout[i];
 
-			fprintf(stderr, "  %d:%15.7g %15.7g %15.7g\n", i, p->x, p->y,
+			fprintf(stderr, "  %d:%15.7g %15.7g %15.7g\n", (int) i, p->x, p->y,
 					p->z);
 		}
 	}
